A whole bunch of futzing with attempting to analyze the ORI image table that starts at 0x130fc
A length field is present at 0x130fc + 2, which stores the length of this table.  It
appears to be 1000*nshots long.  (e.g. 3000 for 3-shot, and 6000 for 6-shot)

This code can now print the offset into that table where a pointer to the start of a JPEG file is found.

Right now, it can find ALL entries in the table for both 3-shot and 6-shot ORIs except for one,
that one is likely the "extra" image already identified by shot counts (e.g. 3-shot ORIs have
25*3*2 + 1 = 151 images, 6-shot have 25*6*2+1 = 301 images)